Refactor api to user app
