--- backend/api/managers.py
@@ -1,20 +1,0 @@
-from django.contrib.auth.base_user import BaseUserManager
-
-class AccountManager(BaseUserManager):
-    def create_user(self, email, password, **extra_fields):
-        if not email:
-            raise ValueError('Email is required')
-        
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, email, password, **extra_fields):
-        extra_fields.update({
-            'is_staff': True,
-            'is_superuser': True,
-        })
-        
-        return self.create_user(email, password, **extra_fields)
