--- backend/api/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from api import views
-
-urlpatterns = [
-    
-]
