+++ backend/backend/urls.py
@@ -18,8 +18,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('api.urls')),
     path('api-auth/', include('rest_framework.urls')),
-    path('rest-auth/', include('rest_auth.urls')),
-    path('rest-auth/registration/', include('rest_auth.registration.urls')),
+    path('accounts/', include('rest_auth.urls')),
+    path('accounts/registration/', include('rest_auth.registration.urls')),
 ]
