+++ backend/users/models.py
@@ -1,8 +1,28 @@
 from django.db import models
+from django.contrib.auth.base_user import BaseUserManager
 from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
-from api.managers import AccountManager
 
-# Create your models here.
+
+class AccountManager(BaseUserManager):
+    def create_user(self, email, password, **extra_fields):
+        if not email:
+            raise ValueError('Email is required')
+        
+        email = self.normalize_email(email)
+        user = self.model(email=email, **extra_fields)
+        user.set_password(password)
+        user.save()
+        return user
+
+    def create_superuser(self, email, password, **extra_fields):
+        extra_fields.update({
+            'is_staff': True,
+            'is_superuser': True,
+        })
+        
+        return self.create_user(email, password, **extra_fields)
+
+
 class UserAccount(AbstractBaseUser, PermissionsMixin):
     display_name = models.CharField('Display Name', max_length=225)
     username = models.CharField(max_length=225)
